Put activePower labels on parallelLines
This is not perfect, because this only sets the lng/lat of the label,
but in the shader the parallel line is actually doing more than just
changing the lng/lat, it also applies min/max distance in pixels. Since
this depends on the zoom level, we don't recompute it in for the labels.
This is not a big problem because we hide the labels when zoomed out,
(min of distance in pixels), and we don't zoom very often past the point where
the lines are at the max number of pixels and the labels move out of the line.

Signed-off-by: Jon Harper <jon.harper87@gmail.com>