increase map maxZoom 12->14
The new version of deckgl/react-map-gl fixes the bug where you could zoom in
more than the max zoom. So now we realize that users actually need a higher
maxZoom.

We had chosen 12 because this is the limit that triggers the bug where
the text label of active power on the map moves away from the arrow on the line
so this will give this bug a little more visibility but it's not a blocker.