Fix map not loading sometimes (#2697)
fix(NetworkMapTab): Sometimes, the tree is loaded quicker than the network map tab component. Then isNetworkModificationTreeUpToDate is true and doesn't changes later. This avoid a necessary proc of the effect which is necessary to call updateMapEquipmentsAndGeoData. If the effect is called only once then previousCurrentRootNetworkUuid is null and we never initialize geodata and mapEquipments

Signed-off-by: sBouzols <sylvain.bouzols@gmail.com>